Break out the SystemInfoScreen from DevScreens.cpp
