+++ UI/DevScreens.h
@@ -99,33 +99,6 @@ private:
 	void OnCompleted(DialogResult result) override;
 };
 
-class SystemInfoScreen : public TabbedUIDialogScreenWithGameBackground {
-public:
-	SystemInfoScreen(const Path &filename) : TabbedUIDialogScreenWithGameBackground(filename) {}
-
-	const char *tag() const override { return "SystemInfo"; }
-
-	void CreateTabs() override;
-	void update() override;
-	void resized() override { RecreateViews(); }
-
-protected:
-	void CopySummaryToClipboard(UI::EventParams &e);
-	bool ShowSearchControls() const override { return false; }
-
-private:
-	void CreateDeviceInfoTab(UI::LinearLayout *deviceInfo);
-	void CreateStorageTab(UI::LinearLayout *storage);
-	void CreateBuildConfigTab(UI::LinearLayout *storage);
-	void CreateCPUExtensionsTab(UI::LinearLayout *storage);
-	void CreateDriverBugsTab(UI::LinearLayout *storage);
-	void CreateInternalsTab(UI::ViewGroup *internals);
-	void CreateOpenGLExtsTab(UI::LinearLayout *gpuExtensions);
-	void CreateVulkanExtsTab(UI::LinearLayout *gpuExtensions);
-
-	int testSliderValue_ = 0;
-};
-
 class GPIGPOScreen : public PopupScreen {
 public:
 	GPIGPOScreen(std::string_view title) : PopupScreen(title, "OK") {}
