+++ UI/UI.vcxproj
@@ -66,6 +66,7 @@
     <ClCompile Include="SavedataScreen.cpp" />
     <ClCompile Include="Store.cpp" />
     <ClCompile Include="AudioCommon.cpp" />
+    <ClCompile Include="SystemInfoScreen.cpp" />
     <ClCompile Include="TabbedDialogScreen.cpp" />
     <ClCompile Include="TiltAnalogSettingsScreen.cpp" />
     <ClCompile Include="TouchControlLayoutScreen.cpp" />
@@ -113,6 +114,7 @@
     <ClInclude Include="RetroAchievementScreens.h" />
     <ClInclude Include="SavedataScreen.h" />
     <ClInclude Include="Store.h" />
+    <ClInclude Include="SystemInfoScreen.h" />
     <ClInclude Include="TabbedDialogScreen.h" />
     <ClInclude Include="TiltAnalogSettingsScreen.h" />
     <ClInclude Include="TouchControlLayoutScreen.h" />
