+++ UI/UI.vcxproj.filters
@@ -123,6 +123,9 @@
       <Filter>Screens</Filter>
     </ClCompile>
     <ClCompile Include="UIAtlas.cpp" />
+    <ClCompile Include="SystemInfoScreen.cpp">
+      <Filter>Screens</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="GameInfoCache.h" />
@@ -246,6 +249,9 @@
       <Filter>Screens</Filter>
     </ClInclude>
     <ClInclude Include="UIAtlas.h" />
+    <ClInclude Include="SystemInfoScreen.h">
+      <Filter>Screens</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <Filter Include="Screens">
