+++ UWP/UI_UWP/UI_UWP.vcxproj.filters
@@ -60,6 +60,7 @@
     </ClCompile>
     <ClCompile Include="..\..\UI\DeveloperToolsScreen.cpp" />
     <ClCompile Include="..\..\UI\UIAtlas.cpp" />
+    <ClCompile Include="..\..\UI\SystemInfoScreen.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="pch.h" />
@@ -121,6 +122,7 @@
     </ClInclude>
     <ClInclude Include="..\..\UI\DeveloperToolsScreen.h" />
     <ClInclude Include="..\..\UI\UIAtlas.h" />
+    <ClInclude Include="..\..\UI\SystemInfoScreen.h" />
   </ItemGroup>
   <ItemGroup>
     <Filter Include="ImDebugger">
